Refactoring and performance improvements of Diffusion model
